requirements.txt
torch
torchvision
fastapi
uvicorn
onnxruntime
pillow
python-multipart